+++ src/listeners/dashboard/MediaListener.ts
@@ -36,8 +36,12 @@ export class MediaListener extends BaseListener {
       case DashboardMessage.revealMedia:
         this.openFileInFinder(msg?.data?.file);
         break;
-      case DashboardMessage.insertPreviewImage:
+      case DashboardMessage.insertMedia:
         Telemetry.send(TelemetryEvent.insertMediaToContent);
+        MediaHelpers.insertMediaToMarkdown(msg?.data);
+        break;
+      case DashboardMessage.insertFile:
+        Telemetry.send(TelemetryEvent.insertFileToContent);
         MediaHelpers.insertMediaToMarkdown(msg?.data);
         break;
       case DashboardMessage.updateMediaMetadata: