+++ src/listeners/dashboard/SettingsListener.ts
@@ -1,4 +1,7 @@
-import { SETTING_CONTENT_STATIC_FOLDER, SETTING_FRAMEWORK_ID } from "../../constants";
+import { join } from "path";
+import { commands, Uri } from "vscode";
+import { Folders } from "../../commands/Folders";
+import { COMMAND_NAME, SETTING_CONTENT_STATIC_FOLDER, SETTING_FRAMEWORK_ID } from "../../constants";
 import { DashboardCommand } from "../../dashboardWebView/DashboardCommand";
 import { DashboardMessage } from "../../dashboardWebView/DashboardMessage";
 import { DashboardSettings, Settings } from "../../helpers";
@@ -25,6 +28,9 @@ export class SettingsListener extends BaseListener {
         break;
       case DashboardMessage.setFramework:
         this.setFramework(msg?.data);
+        break;
+      case DashboardMessage.addFolder:
+        this.addFolder(msg?.data);
         break;
     }
   }
@@ -53,7 +59,7 @@ export class SettingsListener extends BaseListener {
    * Set the current site-generator or framework + related settings
    * @param frameworkId 
    */
-  private static setFramework(frameworkId: string | null) {
+  public static setFramework(frameworkId: string | null) {
     Settings.update(SETTING_FRAMEWORK_ID, frameworkId, true);
 
     if (frameworkId) {
@@ -64,6 +70,16 @@ export class SettingsListener extends BaseListener {
       } else {
         Settings.update(SETTING_CONTENT_STATIC_FOLDER, "", true);
       }
+    }
+
+    SettingsListener.getSettings();
+  }
+
+  private static addFolder(folder: string) {
+    if (folder) {
+      const wsFolder = Folders.getWorkspaceFolder();
+      const folderUri = Uri.file(join(wsFolder?.fsPath || "", folder));
+      commands.executeCommand(COMMAND_NAME.registerFolder, folderUri);
     }
   }
 }