+++ src/panelWebView/components/Fields/TextField.tsx
@@ -6,12 +6,15 @@ export interface ITextFieldProps {
   label: string;
   value: string | null;
   singleLine: boolean | undefined;
+  wysiwyg: boolean | undefined;
   limit: number | undefined;
   rows?: number;
   onChange: (txtValue: string) => void;
 }
 
-export const TextField: React.FunctionComponent<ITextFieldProps> = ({singleLine, limit, label, value, rows, onChange}: React.PropsWithChildren<ITextFieldProps>) => {
+const WysiwygField = React.lazy(() => import('./WysiwygField'));
+
+export const TextField: React.FunctionComponent<ITextFieldProps> = ({singleLine, wysiwyg, limit, label, value, rows, onChange}: React.PropsWithChildren<ITextFieldProps>) => {
   const [ text, setText ] = React.useState<string | null>(value);
 
   React.useEffect(() => {
@@ -37,25 +40,31 @@ export const TextField: React.FunctionComponent<ITextFieldProps> = ({singleLine,
           <PencilIcon style={{ width: "16px", height: "16px" }} /> <span style={{ lineHeight: "16px"}}>{label}</span>
         </div>
       </VsLabel>
-      
+
       {
-        singleLine ? (
-          <input 
-            className={`metadata_field__input`}
-            value={text || ""}
-            onChange={(e) => onTextChange(e.currentTarget.value)} 
-            style={{
-              border: isValid ? "1px solid var(--vscode-inputValidation-infoBorder)" : "1px solid var(--vscode-inputValidation-warningBorder)"
-            }} />
+        wysiwyg ? (
+          <React.Suspense fallback={(<div>Loading field</div>)}>
+            <WysiwygField text={text || ""} onChange={onTextChange} />
+          </React.Suspense>
         ) : (
-          <textarea
-            className={`metadata_field__textarea`}
-            rows={rows || 2}
-            value={text || ""}  
-            onChange={(e) => onTextChange(e.currentTarget.value)} 
-            style={{
-              border: isValid ? "1px solid var(--vscode-inputValidation-infoBorder)" : "1px solid var(--vscode-inputValidation-warningBorder)"
-            }} />
+          singleLine ? (
+            <input 
+              className={`metadata_field__input`}
+              value={text || ""}
+              onChange={(e) => onTextChange(e.currentTarget.value)} 
+              style={{
+                border: isValid ? "1px solid var(--vscode-inputValidation-infoBorder)" : "1px solid var(--vscode-inputValidation-warningBorder)"
+              }} />
+          ) : (
+            <textarea
+              className={`metadata_field__textarea`}
+              rows={rows || 2}
+              value={text || ""}  
+              onChange={(e) => onTextChange(e.currentTarget.value)} 
+              style={{
+                border: isValid ? "1px solid var(--vscode-inputValidation-infoBorder)" : "1px solid var(--vscode-inputValidation-warningBorder)"
+              }} />
+          )
         )
       }