+++ src/panelWebView/components/Metadata.tsx
@@ -4,12 +4,12 @@ import { CommandToCode } from '../CommandToCode';
 import { MessageHelper } from '../../helpers/MessageHelper';
 import { TagType } from '../TagType';
 import { Collapsible } from './Collapsible';
-import { SymbolKeywordIcon } from './Icons/SymbolKeywordIcon';
-import { TagPicker } from './TagPicker';
 import "react-datepicker/dist/react-datepicker.css";
 import useContentType from '../../hooks/useContentType';
-import FieldBoundary from './ErrorBoundary/FieldBoundary';
 import { WrapperField } from './Fields/WrapperField';
+import { ContentTypeValidator } from './ContentType/ContentTypeValidator';
+import { FeatureFlag } from '../../components/features/FeatureFlag';
+import { FEATURE_FLAG } from '../../constants';
 
 export interface IMetadata {
   [prop: string]: string[] | string | null | IMetadata;
@@ -18,10 +18,11 @@ export interface IMetadataProps {
   settings: PanelSettings | undefined;
   metadata: IMetadata;
   focusElm: TagType | null;
+  features: string[];
   unsetFocus: () => void;
 }
 
-const Metadata: React.FunctionComponent<IMetadataProps> = ({settings, metadata, focusElm, unsetFocus}: React.PropsWithChildren<IMetadataProps>) => {
+const Metadata: React.FunctionComponent<IMetadataProps> = ({settings, features, metadata, focusElm, unsetFocus}: React.PropsWithChildren<IMetadataProps>) => {
   const contentType = useContentType(settings, metadata);
 
   const sendUpdate = (field: string | undefined, value: any, parents: string[]) => {
@@ -80,6 +81,12 @@ const Metadata: React.FunctionComponent<IMetadataProps> = ({settings, metadata, 
 
   return (
     <Collapsible id={`tags`} title="Metadata" className={`inherit z-20`}>
+      <FeatureFlag features={features || []} flag={FEATURE_FLAG.panel.contentType}>
+        <ContentTypeValidator
+          fields={contentType?.fields || []}
+          metadata={metadata} />
+      </FeatureFlag>
+
       {
         renderFields(contentType?.fields || [], metadata)
       }