+++ src/panelWebView/components/Tags.tsx
@@ -28,12 +28,24 @@ const Tags: React.FunctionComponent<ITagsProps> = (props: React.PropsWithChildre
     <div className={`article__tags__items`}>
       {
         knownTags.map((t, idx) => (
-          <Tag key={generateKey(t, idx)} value={t} className={`article__tags__items__pill_exists`} onRemove={onRemove} title={`Remove ${t}`} />
+          <Tag 
+            key={generateKey(t, idx)} 
+            value={t} 
+            className={`article__tags__items__pill_exists`} 
+            onRemove={onRemove} 
+            title={`Remove ${t}`} />
         ))
       }
       {
         unknownTags.map((t, idx) => (
-          <Tag key={generateKey(t, idx)} value={t} className={`article__tags__items__pill_notexists`} onRemove={onRemove} onCreate={onCreate} title={`Be aware, this tag "${t}" is not saved in your settings. Once removed, it will be gone forever.`} disableConfigurable={disableConfigurable} />
+          <Tag 
+            key={generateKey(t, idx)} 
+            value={t} 
+            className={`article__tags__items__pill_notexists`} 
+            onRemove={onRemove} 
+            onCreate={onCreate} 
+            title={`Be aware, this tag "${t}" is not saved in your settings. Once removed, it will be gone forever.`} 
+            disableConfigurable={disableConfigurable} />
         ))
       }
     </div>